requirements.txt
dns.resolver
tqdm
termcolor